create_requirements.txt

**Create `requirements.txt` in root:**
```txt
# Backend Requirements
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
alembic==1.12.1
pydantic==2.5.0
pydantic-settings==2.1.0

# Frontend Requirements
streamlit==1.28.1
plotly==5.18.0
pandas==2.1.4
requests==2.31.0

# Common
python-dotenv==1.0.0